rename slices
