+++ client/src/commons/searchbar/SearchBar.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import {useNavigate} from "react-router-dom";
 import { useDispatch } from "react-redux";
-import { searchList, updateSearchWords } from "../../redux/Slices/ProductList.slice";
+import { searchList, updateSearchWords } from "../../redux/Slices/ProductListSlice";
