+++ client/src/components/create/createPhones.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { getProductList } from '../../redux/Slices/ProductList.slice'
+import { getProductList } from '../../redux/Slices/ProductListSlice'
 import { useDispatch, useSelector} from 'react-redux'
 import {useEffect, useState} from 'react'
-import { createProducts } from '../../redux/Slices/ProductList.slice'
+import { createProducts } from '../../redux/Slices/ProductListSlice'
 import { useNavigate } from 'react-router-dom'
 
 function CreatePhones() {
