+++ client/src/pages/Products/ProductsPage.jsx
@@ -1,6 +1,6 @@
 import Products_List from "../../components/Products_List/Products_List";
 import { useDispatch,useSelector } from "react-redux";
-import { getProductList } from "../../redux/Slices/ProductList.slice";
+import { getProductList } from "../../redux/Slices/ProductListSlice";
 
 
 const ProductsPage =()=>{
